backend/requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.9.5
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.7.0
attrs==24.3.0
beautifulsoup4==4.12.3
bidict==0.23.1
blinker==1.9.0
bs4==0.0.2
certifi==2024.12.14
charset-normalizer==3.4.0
click==8.1.7
cohere==5.13.3
colorama==0.4.6
dataclasses-json==0.6.7
distro==1.9.0
fastavro==1.9.7
filelock==3.16.1
Flask==3.1.0
Flask-Cors==5.0.0
Flask-SocketIO==5.4.1
frozenlist==1.5.0
fsspec==2024.10.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.27.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
jiter==0.8.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.12
langchain-cohere==0.3.3
langchain-community==0.3.12
langchain-core==0.3.25
langchain-experimental==0.3.3
langchain-openai==0.2.12
langchain-pinecone==0.2.0
langchain-text-splitters==0.3.3
langsmith==0.2.3
MarkupSafe==3.0.2
marshmallow==3.23.1
multidict==6.1.0
mypy-extensions==1.0.0
numpy==1.26.4
openai==1.57.4
orjson==3.10.12
packaging==24.2
pandas==2.2.3
parameterized==0.9.0
pinecone-client==5.0.1
pinecone-plugin-inference==1.1.0
pinecone-plugin-interface==0.0.7
propcache==0.2.1
pydantic==2.10.3
pydantic-settings==2.7.0
pydantic_core==2.27.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-engineio==4.10.1
python-socketio==5.11.4
pytz==2024.2
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
simple-websocket==1.1.0
six==1.17.0
sniffio==1.3.1
soupsieve==2.6
SQLAlchemy==2.0.36
tabulate==0.9.0
tenacity==9.0.0
tiktoken==0.8.0
tokenizers==0.21.0
tqdm==4.67.1
types-requests==2.32.0.20241016
typing-inspect==0.9.0
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.2.3
Werkzeug==3.1.3
wsproto==1.2.0
yarl==1.18.3
